+++ CHANGELOG.md
@@ -11,7 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add Windows backend #99
 - Add pager support on Windows
-- Add Intel RSD Network card information #81
+- Add Intel RSD Network card information decoder #81
+- Add Intel RSD PCIe information decoder #82
 
 ### Fixed